Support kibana Pvl private link operation.

+++ aliyun-java-sdk-elasticsearch/src/main/java/com/aliyuncs/elasticsearch/model/v20170613/ListActionRecordsRequest.java
@@ -35,11 +35,11 @@ public class ListActionRecordsRequest extends RoaAcsRequest<ListActionRecordsRes
 
 	private String requestId;
 
-	private Integer endTime;
+	private Long endTime;
 
 	private Integer page;
 
-	private Integer startTime;
+	private Long startTime;
 
 	private String userId;
 	public ListActionRecordsRequest() {
@@ -107,11 +107,11 @@ public class ListActionRecordsRequest extends RoaAcsRequest<ListActionRecordsRes
 		}
 	}
 
-	public Integer getEndTime() {
+	public Long getEndTime() {
 		return this.endTime;
 	}
 
-	public void setEndTime(Integer endTime) {
+	public void setEndTime(Long endTime) {
 		this.endTime = endTime;
 		if(endTime != null){
 			putQueryParameter("endTime", endTime.toString());
@@ -129,11 +129,11 @@ public class ListActionRecordsRequest extends RoaAcsRequest<ListActionRecordsRes
 		}
 	}
 
-	public Integer getStartTime() {
+	public Long getStartTime() {
 		return this.startTime;
 	}
 
-	public void setStartTime(Integer startTime) {
+	public void setStartTime(Long startTime) {
 		this.startTime = startTime;
 		if(startTime != null){
 			putQueryParameter("startTime", startTime.toString());
