+++ aliyun-java-sdk-elasticsearch/src/main/java/com/aliyuncs/elasticsearch/model/v20170613/ListLogstashLogRequest.java
@@ -31,9 +31,9 @@ public class ListLogstashLogRequest extends RoaAcsRequest<ListLogstashLogRespons
 
 	private String query;
 
-	private Integer endTime;
+	private Long endTime;
 
-	private Integer beginTime;
+	private Long beginTime;
 
 	private Integer page;
 
@@ -81,22 +81,22 @@ public class ListLogstashLogRequest extends RoaAcsRequest<ListLogstashLogRespons
 		}
 	}
 
-	public Integer getEndTime() {
+	public Long getEndTime() {
 		return this.endTime;
 	}
 
-	public void setEndTime(Integer endTime) {
+	public void setEndTime(Long endTime) {
 		this.endTime = endTime;
 		if(endTime != null){
 			putQueryParameter("endTime", endTime.toString());
 		}
 	}
 
-	public Integer getBeginTime() {
+	public Long getBeginTime() {
 		return this.beginTime;
 	}
 
-	public void setBeginTime(Integer beginTime) {
+	public void setBeginTime(Long beginTime) {
 		this.beginTime = beginTime;
 		if(beginTime != null){
 			putQueryParameter("beginTime", beginTime.toString());
